wrapping components in storybook with h-fit
