+++ lexio/lib/components/SourcesDisplay/SourcesDisplay.tsx
@@ -39,31 +39,26 @@ export interface SourcesDisplayProps {
    * @default "Retrieved Sources"
    */
   title?: string;
-
   /**
    * Placeholder text for the search input field
    * @default "Search sources..."
    */
   searchPlaceholder?: string;
-
   /**
    * Controls visibility of the search functionality
    * @default true
    */
   showSearch?: boolean;
-
   /**
    * Controls visibility of relevance scores
    * @default true
    */
   showRelevanceScore?: boolean;
-
   /**
    * Controls visibility of metadata tags
    * @default true
    */
   showMetadata?: boolean;
-
   /**
    * Style customization options
    */
