+++ lexio/src/stories/components/HTMLViewer.stories.tsx
@@ -2,10 +2,9 @@ import type { Meta, StoryObj } from '@storybook/react';
 import 'tailwindcss/tailwind.css';
 import {RAGProvider} from "../../../lib/components/RAGProvider";
 import {HtmlViewer} from "../../../lib/components/Viewers";
-import {extractComponentDescription} from "@storybook/docs-tools";
+import {extractComponentDescriptionHelper} from "./helper.tsx"
 
-const testContent = `
-<div style="font-family: Arial, sans-serif; padding: 20px; line-height: 1.6; color: #333;">
+const testContent = `<div style="font-family: Arial, sans-serif; padding: 20px; line-height: 1.6; color: #333;">
   <h1 style="color: #4A90E2;">Test HTML Content</h1>
   <p>This is a <strong>test paragraph</strong> to showcase HTML rendering in the <code>HtmlViewer</code> component.</p>
 
@@ -43,32 +42,12 @@ const meta: Meta<typeof HtmlViewer> = {
   tags: ['autodocs'],
   parameters: {
         docs: {
-            extractComponentDescription: (component, { notes }) => {
-                // Use Storybook's default extractor
-                let description = extractComponentDescription(component, { notes });
-
-                if (description) {
-                    // Customize the description by filtering unwanted TSDoc decorators
-                    description = description
-                        .split('\n') // Split lines
-                        .filter(line =>
-                            !line.startsWith('@param') &&  // Remove @param tags
-                            !line.startsWith('@component') &&  // Remove @param tags
-                            !line.startsWith('@remarks') &&  // Remove @param tags
-                            !line.startsWith('@example') &&  // Remove @param tags
-                            !line.startsWith('@todo') &&  // Remove @param tags
-                            !line.startsWith('@returns')   // Remove @returns tags
-                        )
-                        .join('\n'); // Rejoin the lines
-                }
-
-                return description || ""; // Return modified description
-            },
+            extractComponentDescription: extractComponentDescriptionHelper
         },
     },
   decorators: [
     (Story) => (
-      <div style={{ width: '600px', minHeight: '600px', padding: '1rem' }}>
+      <div className="h-fit" style={{ width: '600px', padding: '1rem' }}>
         <RAGProvider>
           <Story />
           </RAGProvider>
