+++ lexio/src/stories/components/PDFViewer.stories.tsx
@@ -1,9 +1,9 @@
 import type {Meta, StoryObj} from '@storybook/react';
-import { extractComponentDescription } from '@storybook/docs-tools';
 import React, {useEffect, useState} from 'react';
 import {PdfViewer} from '../../../lib/components/Viewers/PdfViewer';
 import 'tailwindcss/tailwind.css';
 import {RAGProvider} from "../../../lib/components/RAGProvider";
+import {extractComponentDescriptionHelper} from "./helper.tsx"
 
 async function fetchPdfAsUint8Array(url: string): Promise<Uint8Array> {
     try {
@@ -47,32 +47,12 @@ const meta: Meta<typeof PdfViewer> = {
     tags: ['autodocs'],
     parameters: {
         docs: {
-            extractComponentDescription: (component, { notes }) => {
-                // Use Storybook's default extractor
-                let description = extractComponentDescription(component, { notes });
-
-                if (description) {
-                    // Customize the description by filtering unwanted TSDoc decorators
-                    description = description
-                        .split('\n') // Split lines
-                        .filter(line =>
-                            !line.startsWith('@param') &&  // Remove @param tags
-                            !line.startsWith('@component') &&  // Remove @param tags
-                            !line.startsWith('@remarks') &&  // Remove @param tags
-                            !line.startsWith('@example') &&  // Remove @param tags
-                            !line.startsWith('@todo') &&  // Remove @param tags
-                            !line.startsWith('@returns')   // Remove @returns tags
-                        )
-                        .join('\n'); // Rejoin the lines
-                }
-
-                return description || ""; // Return modified description
-            },
+            extractComponentDescription: extractComponentDescriptionHelper,
         },
     },
     decorators: [
         (Story) => (
-            <div style={{width: '600px', height: '600px', padding: '1rem'}}>
+            <div className="h-[600px]"  style={{width: '600px', padding: '1rem'}}>
                 <RAGProvider>
                     <Story/>
                 </RAGProvider>
